+++ lib/src/analyzers/lint_analyzer/lint_analyzer.dart
@@ -70,9 +70,11 @@ class LintAnalyzer {
   Future<Iterable<LintFileReport>> runCliAnalysis(
     Iterable<String> folders,
     String rootFolder,
-    LintConfig config,
-  ) async {
-    final collection = createAnalysisContextCollection(folders, rootFolder);
+    LintConfig config, [
+    String? sdkPath,
+  ]) async {
+    final collection =
+        createAnalysisContextCollection(folders, rootFolder, sdkPath);
 
     final analyzerResult = <LintFileReport>[];