+++ lib/src/cli/commands/check_unused_files.dart
@@ -27,6 +27,7 @@ class CheckUnusedFilesCommand extends BaseCommand {
   @override
   void validateCommand() {
     validateRootFolderExist();
+    validateSdkPath();
     validateTargetDirectories();
   }
 
@@ -43,6 +44,7 @@ class CheckUnusedFilesCommand extends BaseCommand {
       folders,
       rootFolder,
       config,
+      findSdkPath(),
     );
 
     await _analyzer
@@ -56,6 +58,7 @@ class CheckUnusedFilesCommand extends BaseCommand {
   void _addFlags() {
     _usesReporterOption();
     usesRootFolderOption();
+    usesSdkPathOption();
     usesExcludeOption();
   }