+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * feat: introduce summary report.
 * chore: deprecate documentation in Github repo.
 * chore: restrict `analyzer` version to `>=2.4.0 <2.7.0`.
+* CLI now can be compiled to and used as Windows executable.
 
 ## 4.5.0-dev.1